String changes
"Google Now On Tap" is a thing of the past. These days especially on all
officially supported devices, Google assistant is what comes up when you
use this. Changed to Google assistant.

'Settings panel' is a bit misleading. When I saw this at first I assumed
it had something to do with the acutal Settings app. Lets leave zero doubt
what this is by changing it to QS panel.

Change-Id: Ic6559a2c6be65484cde9b35fe50de6664979a3a3