Fix FC for Fling when unpinning an app
to replicate: enable fling navbar, pin an app, stock statusbar will come up.
then long press back and recent buttons to unpin the app, but keep
tapping with the fingers a bit longer then release the tap.

This is because when after the unpin event we switch back to Fling
navbar, FlingView onAttachedToWindow sets the mFlingTouchListener
that sends the event to the FlingGestureHandler onFling, but the start
event is null so SystemUI FC.

So we can just add a null check there to fix the issue.

Change-Id: I08b37054e819b174bdc6ba19a0db3ea34822461d